requirements.txt
# requirements.txt

# For the "langchain_google_genai" integration with Gemini
langchain_google_genai==0.1.3

# This library typically requires google-generativeai
google-generativeai==0.3.2

# For general LangChain usage (sometimes needed for schema, etc.)
langchain==0.1.1

# The LangGraph library for multi-agent state graphs
langgraph==0.2.0

# For environment variable loading
python-dotenv==1.0.0

# For the web UI
streamlit==1.29.0

# For typed dictionaries (backport in some older Python versions)
typing_extensions==4.9.0